Refactor with ExportStats component
