+++ packages/lib-user/src/components/Contributors/components/ExportStats/index.js
@@ -1,0 +1,1 @@
+export { default } from './ExportStats'
